Result analysis
As you will see on upcomming slides, it is difficult to find a clear link between the price of accomodation, the number of reviews, its availability or the required minimum nights.

To sum up, Berlin has four most popular districts where you can find a comfortable accommodation using Airbnb.com. If you want a cheap room, it is best to look in Mitte. Most places with average prices can be found in Pankow. If you are interested in very expensive objects, I recommend Tempelhof-Schoneberg and Neukolln. Most offers are up to 69 EUR per night. We can choose between renting entire or apartments and renting private or shared rooms. Most of the facilities have few reviews and require a short stay of up to 3 days. Some facilities are available up to 365 days a year. 

Entire homes, which are characterized by high availability, suggest that the owner does not live there, and that it might be regular rental business for him such as renting facilities for tourists, which is illegal in some cities (without registered business activity).
In March 2018 Berlin formally allowed vacation rentals again, but there’s a strict new set of rules—and even tougher penalties.
Starting May 1, owner-occupiers will, under certain conditions, be allowed once more to rent out their own home as much as they want, and to rent out second homes for up to 90 days a year.